### 2025-04-30 #8140 @uislambekov
Potential reviewers:
Name: Matt Piekenbrock; github: @peekxc
Name: @EduPH; github: @EduPH

On Wed, Apr 30, 2025 at 2:22 PM The Open Journals editorial robot <
***@***.***> wrote:

> *editorialbot* created an issue (openjournals/joss-reviews#8140)
> <https://github.com/openjournals/joss-reviews/issues/8140>
>
> *Submitting author:* @uislambekov <https://github.com/uislambekov> (Umar
> Islambekov <http://orcid.org/0000-0002-5103-3238>)
> *Repository:* https://github.com/uislambekov/TDAvec
> *Branch with paper.md* (empty if default branch): JOSS
> *Version:* 0.1.41
> *Editor:* Pending
> *Reviewers:* Pending
> *Managing EiC:* Chris Vernon
> Status
>
> [image: status]
> <https://joss.theoj.org/papers/3c986e2111e0a64c76a56442e31a62e6>
>
> Status badge code:
>
>
> *Author instructions*
>
> Thanks for submitting your paper to JOSS @uislambekov
> <https://github.com/uislambekov>. *Currently, there isn't a JOSS editor
> assigned* to your paper.
>
> @uislambekov <https://github.com/uislambekov> if you have any suggestions
> for potential reviewers then please mention them here in this thread
> (without tagging them with an @). You can search the list of people
> <https://reviewers.joss.theoj.org/lookup> that have already agreed to
> review and may be suitable for this submission.
>
> *Editor instructions*
>
> The JOSS submission bot @editorialbot <https://github.com/editorialbot>
> is here to help you find and assign reviewers and start the main review. To
> find out what @editorialbot <https://github.com/editorialbot> can do for
> you type:
>
> @editorialbot commands
>
