### 2025-04-30 #8140 @uislambekov
On Wed, Apr 30, 2025 at 2:22 PM The Open Journals editorial robot <
***@***.***> wrote:

> *editorialbot* left a comment (openjournals/joss-reviews#8140)
> <https://github.com/openjournals/joss-reviews/issues/8140#issuecomment-2842908834>
>
> Hello human, I'm @editorialbot <https://github.com/editorialbot>, a robot
> that can help you with some common editorial tasks.
>
> For a list of things I can do to help you, just type:
>
> @editorialbot commands
>
> For example, to regenerate the paper pdf after making changes in the
> paper's md or bib files, type:
>
> @editorialbot generate pdf
>
