+++ JOSS/paper.md
@@ -10,7 +10,7 @@ authors:
   email: aluchi@bgsu.edu
   affiliation: 1
   corresponding: yes
-date: "2025-04-14"
+date: "2025-06-16"
 output: pdf_document
 bibliography: paper.bib
 tags:
@@ -32,7 +32,7 @@ A topological descriptor outputted by the persistent homology encoding the shape
 
 # Statement of need
 
-The problem of transforming PDs into finite dimensional vectors for machine learning purposes has attracted considerable attention in the TDA research community over the past decade. Early vector summaries of PDs - such as the persistence landscape [@bubenik2015statistical], persistence silhouette [@chazal2014stochastic], Betti curve\footnote{Also known as the Betti function} [@chazal2021introduction], and persistence image [@adams2017persistence] - have been implemented in both \texttt{Python} and \texttt{R} packages. However, there remains a need for a unified package that brings these methods (both the classical approaches and those developed more recently) together using consistent syntax and efficient implementation. The \texttt{TDAvec} package, available in both \texttt{R} and \texttt{Python}, is designed to meet this need. Its contributions can be summarized in the following three areas:  
+The problem of transforming PDs into finite dimensional vectors for machine learning purposes has attracted considerable attention in the TDA research community over the past decade. Early vector summaries of PDs - such as the persistence landscape [@bubenik2015statistical], persistence silhouette [@chazal2014stochastic], Betti curve\footnote{Also known as the Betti function} [@chazal2021introduction], and persistence image [@adams2017persistence] - have been implemented in both \texttt{Python} and \texttt{R} packages. However, there remains a need for a unified package that brings these methods (both the classical approaches and those developed more recently) together using consistent syntax and efficient implementation. The \texttt{TDAvec} package, available in both \texttt{R} and \texttt{Python}, is designed to meet this need. Its contributions can be summarized in the following three areas: 
 
 1. It extends the list of implemented vector summaries for PDs by incorporating 13 vectorization methods used in TDA. These methods can be grouped into three broad categories:
 
