### 2025-06-21 #8140 @ujjwalkarn
@fernandosola, @iliapopov17, @adityapt -- would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [https://joss.readthedocs.io/en/latest/review_criteria.htm](https://joss.readthedocs.io/en/latest/review_criteria.html)
