### 2025-07-02 #8532 @editorialbot
[REVIEW]: TDAvec: Computing Vector Summaries of Persistence Diagrams for Topological Data Analysis in R and Python
**Submitting author:** @uislambekov (<a href="http://orcid.org/0000-0002-5103-3238">Umar Islambekov</a>)
**Repository:** https://github.com/uislambekov/TDAvec
**Branch with paper.md** (empty if default branch): main
**Version:** 0.1.41
