### 2025-07-04 uislambekov/TDAvec#7 @adityapt
TDAvec Package Installation Error - on Mac - Rstudio
This is related to the following review: [review #8532](https://github.com/openjournals/joss-reviews/issues/8532 )

I'm using R 4.5.1 in MACBOOK Pro.

The package installation fails when I use devtools::install_github("uislambekov/TDAvec", build_vignettes = TRUE) method. Below is the error:

Installing 2 packages: Rcpp, RcppArmadillo
trying URL 'https://cran.rstudio.com/bin/macosx/big-sur-arm64/contrib/4.5/Rcpp_1.0.14.tgz'
trying URL 'https://cran.rstudio.com/bin/macosx/big-sur-arm64/contrib/4.5/RcppArmadillo_14.4.3-1.tgz'

The downloaded binary packages are in
	/var/folders/yd/lby68ngs697dh0mq13pg1b4h0000gq/T//RtmpbHFrO7/downloaded_packages
── R CMD build ─────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘/private/var/folders/yd/lby68ngs697dh0mq13pg1b4h0000gq/T/RtmpbHFrO7/remotes17bbd3d773ae4/uislambekov-TDAvec-83575fa/DESCRIPTION’ ...
─  preparing ‘TDAvec’:
✔  checking DESCRIPTION meta-information
─  cleaning src
E  creating vignettes (40.1s)
   --- re-building ‘TDAvec_vignette.Rmd’ using rmarkdown

   Quitting from TDAvec_vignette.Rmd:44-47 [setup]
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   <error/rlang_error>
   Error in `library()`:
   ! there is no package called 'TDAstats'
   ---
   Backtrace:
       ▆
    1. └─base::library(TDAstats)
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

   Error: processing vignette 'TDAvec_vignette.Rmd' failed with diagnostics:
   there is no package called 'TDAstats'
   --- failed re-building ‘TDAvec_vignette.Rmd’

   SUMMARY: processing the following file failed:
     ‘TDAvec_vignette.Rmd’

   Error: Vignette re-building failed.
   Execution halted
Error: Failed to install 'TDAvec' from GitHub:
  ! System command 'R' failed
