### 2025-07-04 uislambekov/TDAvec#7 @uislambekov
Try installing TDAstats manually first by running:

install.packages("TDAstats")

Alternatively, you can install TDAvec by skipping building vignettes:

devtools::install_github("uislambekov/TDAvec", build_vignettes = FALSE)

I hope this is helpful.
