### 2025-07-04 uislambekov/TDAvec#7 @adityapt
Thanks! That worked (and I had to do that), but I think to avoid build errors and for the installation to pass, you may want to add this chunk to your Vignette:

### ---- Conditional dependency logic ----
stats_pkg_flag <- requireNamespace("TDAstats", quietly = TRUE)

if (stats_pkg_flag) {
  library(TDAstats)      # load it only if present
} else {
  message("NOTE: TDAstats not installed; code chunks that rely on it will be shown but not run.")
  ### Turn off evaluation for all later chunks that rely on TDAstats
  knitr::opts_chunk$set(eval = FALSE)
}
