### 2025-07-04 uislambekov/TDAvec#7 @corybrunson
Is there a specific reason to do this as opposed to adding `%\VignetteDepends{TDAstats}` to the front matter? (This is in reference to ["Writing R Extensions"](https://cran.r-project.org/doc/manuals/R-exts.html#Writing-package-vignettes-1).)

By the way, {TDAstats} is no longer actively maintained, so {TDA} or {ripserr} would be a safer "soft" dependency here.
