### 2025-07-04 uislambekov/TDAvec#7 @adityapt
Good point @corybrunson. %\VignetteDepends{TDAstats} might still fail if the package is missing - and I notice that the Description file didn't have TDAstats in the requirements, but the logic I gave will silently not throw any build error and go ahead with the installation. The goal is to avoid installation errors I guess.

Also, if TDAstats is not maintained, let's ask the author to consider replacing it in the Vignette or raising a comment somewhere that this package is no longer maintained.
