### 2025-07-04 uislambekov/TDAvec#7 @corybrunson
Oh, i see how that would work, thanks @adityapt. Though it feels dodgy to me and i would urge that packages necessary for the vignette be included in `Suggests:` anyway.

{TDAstats} has few dependencies, none of them at risk of being orphaned, so i don't think it's a pressing issue, but this seemed like a good opportunity to mention it.
