### 2025-07-04 uislambekov/TDAvec#7 @adityapt
Got it @corybrunson. Suggets is actually the right thing to do. But, looks like @uislambekov has added it to the list of dependencies in the DESCRIPTION file and that should fix it. I will go ahead and close this issue.
