### 2025-07-04 uislambekov/TDAvec#7 @uislambekov
I have listed TDAstats under Imports in the DESCRIPTION file. Originally, our vignette used the TDA package to compute persistence diagrams. However, a few months ago, TDA was temporarily unavailable on CRAN, and we received a notification from CRAN regarding this. To ensure the continued availability and functionality of our package, we switched to using TDAstats as an alternative.
