### 2025-07-04 uislambekov/TDAvec#7 @corybrunson
@uislambekov that makes sense. For context, i've observed over the years that {TDA} occasionally gets archived but is quickly restored; that is to say, it is under active development, even if the maintainers cannot patch is as quickly as CRAN requires.
