### 2025-07-12 uislambekov/TDAvec#8 @adityapt
R CRAN Installation Doesn't Have any Unit Tests
This is related to the following review: [issues/8532](https://github.com/openjournals/joss-reviews/issues/8532)

I've been going through the package here in Github as well as CRAN, and I don't see any unit tests written for the R version specifically, apart from the vignettes.

Just wanted to confirm if I am missing anything here. Thanks!

If the unit tests are not setup, it is recommended to setup tests for any package that's deployed through CRAN.
