### 2025-07-13 uislambekov/TDAvec#9 @adityapt
Contribution, Code of Conduct and Support Guidelines - Needed for JOSS Review
This is regarding Community-guidelines requirement (JOSS checklist) -  [#8532](https://github.com/openjournals/joss-reviews/issues/8532)

I am reviewing the structure and currently the repository doesn’t include the following:

CONTRIBUTING.md  – how to set up a dev environment, coding style, tests, and the pull-request process.

CODE_OF_CONDUCT.md  – a behaviour policy (e.g. the standard Contributor Covenant v2.1).

Support / help note  – a sentence in the README (or SUPPORT.md) pointing users to the GitHub Issues page (or Discussions) for questions and bug reports.

Could you please add these files for me to review them? These are essential for the review to be completed.
I was able to find an example for you to get started with:

Contribution file: https://github.com/ashleychontos/pySYD/blob/master/CONTRIBUTING.md
Code of Conduct file: https://github.com/ashleychontos/pySYD/blob/master/CODE_OF_CONDUCT.md

Thanks!
