### 2025-07-15 uislambekov/TDAvec#8 @uislambekov
You're right - unit tests aren't set up for TDAvec. Thank you for pointing that out; we'll work on adding them.
