+++ CONTRIBUTING.md
@@ -1,0 +1,30 @@
+# Contributing to TDAvec
+
+Thank you for considering contributing to **TDAvec**!  
+We welcome contributions of all kinds, including bug reports, feature requests, documentation improvements, and code contributions.
+
+## How you can contribute
+
+### 🐛 Reporting bugs
+
+- Use the [GitHub Issues](https://github.com/uislambekov/TDAvec/issues) page to report bugs.
+- Please include:
+  - A minimal reproducible example (e.g. using `reprex`).
+  - The expected behavior.
+  - The actual behavior.
+  - Your R session info (`sessionInfo()`).
+
+### 💡 Suggesting enhancements
+
+- Open an issue to suggest new features or improvements.
+- Explain the motivation and possible use cases.
+
+### ✍️ Improving documentation
+
+- Fixing typos, clarifying examples, or adding new examples is always welcome.
+- You can submit these directly as a pull request.
+
+### 🛠️ Code contributions (pull requests)
+
+1. Fork the repository and clone your fork locally.
+2. Create a new branch for your fix or feature:
