+++ DESCRIPTION
@@ -15,3 +15,6 @@ LinkingTo: Rcpp, RcppArmadillo
 VignetteBuilder: knitr
 URL: https://github.com/uislambekov/TDAvec
 BugReports: https://github.com/uislambekov/TDAvec/issues
+Suggests: 
+    testthat (>= 3.0.0)
+Config/testthat/edition: 3
