### 2025-07-18 uislambekov/TDAvec#10 @adityapt
References in the Paper have some issues
This is related to the following review: https://github.com/openjournals/joss-reviews/issues/8532

I noticed that some of the references have issues in this: https://github.com/uislambekov/TDAvec/blob/main/paper.bib

Could you work on those and fix it?

<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
<link id=Main-File rel=Main-File
href="file:///C:/Users/putta/AppData/Local/Temp/msohtmlclip1/01/clip.htm">
<link rel=File-List
href="file:///C:/Users/putta/AppData/Local/Temp/msohtmlclip1/01/clip_filelist.xml">
<style>

</style>
</head>

<body link="#467886" vlink="#96607D">

Key (or   first-author & year) | Issue(s) to fix
-- | --
ali 2023 – IEEE TPAMI survey | DOI missing
awesome-tda   2024 (GitHub repo) | No author/organisation field
christopher   2018 – Ripser.py | Incomplete citation &   missing DOI
di Fabio   & Ferri 2015 – ICIAP LNCS chapter | DOI missing
ferri   1999 – Representing size   functions … | Venue name unclear
GUDHI   manual 2020 | No DOI/URL field
kernelTDA (CRAN package) | DOI missing
kalivsnik   2019 – FoCM “Tropical   coordinates …” | DOI missing
kepler   mapper 2019 – JOSS article | DOI missing
perea   2023 – FoCM “Approximating   continuous functions …” | DOI missing
Persim   2019 (website) | Optional: add authors

</body>

</html>
