+++ paper.md
@@ -1,0 +1,69 @@
+---
+title: 'TDAvec: Computing  Vector Summaries of Persistence Diagrams for Topological
+  Data Analysis in R and Python'
+authors:
+- name: Umar Islambekov
+  orcid: "0000-0002-5103-3238"
+  affiliation: 1
+  equal-contrib: yes
+- name: Aleksei Luchinsky
+  email: aluchi@bgsu.edu
+  affiliation: 1
+  corresponding: yes
+date: "2025-06-16"
+output: pdf_document
+bibliography: paper.bib
+tags:
+- Persistent Homology
+- Topological Data Analysis
+- Persistence Diagram
+- Vectorization of Persistence Diagrams
+affiliations:
+- index: 1
+  name: Bowling Green State University, USA
+---
+
+
+# Summary
+
+The theory of \emph{persistent homology} is one of the popular tools in \emph{topological data analysis} (TDA) to analyze data with underlying shape structure [@Carlsson:2009; @edelsbrunner2010computational; @chazal2021introduction]. In this context, a single data observation could be a collection of points lying in a metric space, an image, a graph or a time series. The basic idea behind persistent homology is to build a nested sequence (or \emph{filtration}) of \emph{simplicial complexes} (indexed by a scale parameter) on top of data points and keep a record of the appearance and disappearance of various topological features at different scale values. Here, these topological features are "holes" of different dimensions -- connected components, loops, voids, and their higher-dimensional versions whose emergence and subsequent disappearance are tracked using a concept of homology from algebraic topology. From a geometric point of view, simplicial complexes consist of vertices, edges, triangles, tetrahedra etc., glued together and serve as a means for recovering (at least partially) the underlying shape information which is lost during sampling [@nanda2013simplicial].
+
+A topological descriptor outputted by the persistent homology encoding the shape of data is called a \emph{persistence diagram} (PD). Mathematically, a $k$-dimensional PD is a multi-set of points $D=\{(b_i,d_i)\}_{i=1}^N$, where each point $(b_i,d_i)$ corresponds to a topological feature of homological dimension $k$ (0 if a connected component, 1 if a loop, 2 if a void, etc) with the $x$-coordinate representing the scale at which this feature is born (or created), and the $y$-coordinate representing the scale at which it dies (or disappears). In practice, one is usually interested in applying a machine learning method to PDs to make further inferences from data. However, the fact that PDs do not form a Hilbert space, which is a feature (or an input) space for a wide class of machine learning methods, limits their direct of use in applications. To overcome this challenge, kernel methods and vectorization techniques are commonly used [@chung2022persistence]. The kernel approach involves defining a notion of similarity between pairs of PDs, whereas the vectorization methods aim to transform PDs into finite-dimensional feature vectors that can be used as input for many standard machine learning models.   
+
+# Statement of need
+
[…]
