+++ paper.bib
@@ -1,0 +1,362 @@
+@book{edelsbrunner2010computational,
+  title={Computational topology: an introduction},
+  author={Edelsbrunner, Herbert and Harer, John},
+  year={2010},
+  publisher={American Mathematical Soc.}
+}
+
+@article{bubenik2015statistical,
+  title={Statistical topological data analysis using persistence landscapes},
+  author={Bubenik, Peter},
+  journal={The Journal of Machine Learning Research},
+  volume={16},
+  number={1},
+  pages={77--102},
+  year={2015},
+  publisher={JMLR. org}
+}
+
+@inproceedings{chazal2014stochastic,
+  title={Stochastic convergence of persistence landscapes and silhouettes},
+  author={Chazal, Fr{\'e}d{\'e}ric and Fasy, Brittany Terese and Lecci, Fabrizio and Rinaldo, Alessandro and Wasserman, Larry},
+  booktitle={Proceedings of the thirtieth annual symposium on Computational geometry},
+  pages={474--483},
+  year={2014},
+  doi = {10.1145/2582112.2582128}
+}
+
+@article{adams2017persistence,
+  title={Persistence images: A stable vector representation of persistent homology},
+  author={Adams, Henry and Emerson, Tegan and Kirby, Michael and Neville, Rachel and Peterson, Chris and Shipman, Patrick and Chepushtanova, Sofya and Hanson, Eric and Motta, Francis and Ziegelmeier, Lori},
+  journal={The Journal of Machine Learning Research},
+  volume={18},
+  number={1},
+  pages={218--252},
+  year={2017},
+  publisher={JMLR. org}
+}
+
+@article{Carlsson:2009,
+  title={Topology and Data},
+  author={Carlsson, G.},
+  journal={Bulletin of the American Mathematical Society},
+  volume={46},
+  number={2},
+  year={2009},
+  doi={10.1090/S0273-0979-09-01249-X}
+}
+
+@article{chung2022persistence,
+	title={Persistence curves: A canonical framework for summarizing persistence diagrams},
+	author={Chung, Yu-Min and Lawson, Austin},
+	journal={Advances in Computational Mathematics},
+	volume={48},
+	number={1},
+	pages={1--42},
+	year={2022},
+	publisher={Springer},
+	doi={10.1007/s10444-021-09893-4}
+}
+
+@article{chazal2021introduction,
+	title={An introduction to topological data analysis: fundamental and practical aspects for data scientists},
+	author={Chazal, Fr{\'e}d{\'e}ric and Michel, Bertrand},
+	journal={Frontiers in artificial intelligence},
+	volume={4},
+	year={2021},
+	publisher={Frontiers Media SA},
+	doi={10.3389/frai.2021.667963}
+}
+
+@article{hensel2021survey,
+	title={A survey of topological machine learning methods},
+	author={Hensel, Felix and Moor, Michael and Rieck, Bastian},
+	journal={Frontiers in Artificial Intelligence},
+	volume={4},
+	pages={681108},
+	year={2021},
+	publisher={Frontiers Media SA},
+	doi={10.3389/frai.2021.681108}
+}
+
+@Manual{TDAvec,
+	title = {TDAvec: Vector Summaries of Persistence Diagrams},
+	author = {Umar Islambekov and Alexey Luchinsky},
+	year = {2022},
+	note = {R package version 0.1.2},
+	url = {https://CRAN.R-project.org/package=TDAvec},
+	doi={10.32614/CRAN.package.TDAvec}
+}
+
+@Manual{TDA,
[…]
