+++ JOSS/paper.bib
@@ -109,57 +109,6 @@
 	doi = {10.21105/joss.00860}
 }
 
-@Manual{kernelTDA,
-	title = {kernelTDA: Statistical Learning with Kernel for Persistence Diagrams},
-	author = {Tullia Padellini and Francesco Palini},
-	year = {2020},
-	note = {R package version 1.0.0},
-	url = {https://CRAN.R-project.org/package=kernelTDA}
-}
-
-@Manual{TDAmapper,
-	title = {TDAmapper: Topological Data Analysis using Mapper},
-	author = {Paul Pearson and Daniel Muellner and Gurjeet Singh},
-	year = {2015},
-	note = {R package version 1.0},
-	url = {https://github.com/paultpearson/TDAmapper}
-}
-
-@Manual{TDAkit,
-	title = {TDAkit: Toolkit for Topological Data Analysis},
-	author = {Kisung You and Byeongsu Yu},
-	year = {2021},
-	note = {R package version 0.1.2},
-	url = {https://CRAN.R-project.org/package=TDAkit},
-	doi={10.32614/cran.package.tdakit}
-}
-
-@Manual{tdaunif,
-	title = {tdaunif: Uniform Manifold Samplers for Topological Data Analysis},
-	author = {Jason C. Brunson and Brandon Demkowicz and  Sanmati Choudhary},
-	year = {2024},
-	note = {R package version 0.2.0},
-	url = {https://CRAN.R-project.org/package=tdaunif},
-	doi={10.32614/CRAN.package.tdaunif}
-}
-
-@Manual{TDApplied,
-	title = {TDApplied: Machine Learning and Inference for Topological Data Analysis},
-	author = {Shael Brown and Reza Farivar},
-	year = {2024},
-	note = {R package version 3.0.3},
-	url = {https://CRAN.R-project.org/package=TDApplied},
-	doi={10.32614/cran.package.tdapplied}
-}
-
-@Manual{ripserr,
-	title = {ripserr: Calculate Persistent Homology with Ripser-Based Engines},
-	author = {Raoul Wadhwa and Matt Piekenbrock and Jacob Scott},
-	year = {2020},
-	note = {R package version 0.1.1},
-	url = {https://CRAN.R-project.org/package=ripserr},
-	doi={10.32614/CRAN.package.ripserr}
-}
 
 @article{atienza2020stability,
 	title={On the stability of persistent entropy and new summary functions for topological data analysis},
@@ -221,14 +170,6 @@
 	URL = {https://doi.org/10.1007/s41468-021-00071-5},
 }
 
-  @Book{ggplot2,
-	author = {Hadley Wickham},
-	title = {ggplot2: Elegant Graphics for Data Analysis},
-	publisher = {Springer-Verlag New York},
-	year = {2016},
-	isbn = {978-3-319-24277-4},
-	url = {https://ggplot2.tidyverse.org},
-}
 
   @Manual{Rcpp,
 	title = {Rcpp: Seamless R and C++ Integration},
@@ -241,61 +182,7 @@
 	doi={10.32614/CRAN.package.Rcpp}
 }
 
-@online{awesome-tda:2024,
-	url = {https://github.com/FatemehTarashi/awesome-tda},
-	title        = {Awesome-TDA},
-	year         = {2024},
-	urldate = {2024-08-12}
-}
 
-@article{tauzin2021giotto,
-  title={giotto-tda:: A topological data analysis toolkit for machine learning and data exploration},
-  author={Tauzin, Guillaume and Lupo, Umberto and Tunstall, Lewis and P{\'e}rez, Julian Burella and Caorsi, Matteo and Medina-Mardones, Anibal M and Dassatti, Alberto and Hess, Kathryn},
-  journal={Journal of Machine Learning Research},
-  volume={22},
-  number={39},
[…]
