+++ JOSS/paper.bib
@@ -1,9 +1,15 @@
-@book{edelsbrunner2010computational,
-  title={Computational topology: an introduction},
-  author={Edelsbrunner, Herbert and Harer, John},
-  year={2010},
-  publisher={American Mathematical Soc.}
+@article{edelsbrunner2002topological,
+  title={Topological persistence and simplification},
+  author={Edelsbrunner and Letscher and Zomorodian},
+  journal={Discrete \& computational geometry},
+  volume={28},
+  number={4},
+  pages={511--533},
+  year={2002},
+  publisher={Springer},
+  doi = {10.1007/s00454-002-2885-2}
 }
+
 
 @article{bubenik2015statistical,
   title={Statistical topological data analysis using persistence landscapes},
