### 2025-07-18 uislambekov/TDAvec#10 @uislambekov
Thank you! I've fixed the issues with the references. I've also updated the paper.md file in the root GitHub directory — it’s the same file located in the JOSS folder. I realized that the version in the root directory hadn’t been properly updated. This revision was made in response to the request to the keep word count under 1000.
