### 2025-07-19 uislambekov/TDAvec#10 @adityapt
Thanks! I feel that this is incorrect though:

@article{christopher2018lean,
  title={A lean persistent homology library for python},
  author={Christopher, T and Nathaniel, S and Rann, BO},
  journal={Open J},
  volume={925},
  year={2018}
}

This is what I get from JOSS website:

https://joss.theoj.org/papers/10.21105/joss.00925#

@article{Tralie2018, doi = {10.21105/joss.00925}, url = {https://doi.org/10.21105/joss.00925}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {29}, pages = {925}, author = {Tralie, Christopher and Saul, Nathaniel and Bar-On, Rann}, title = {Ripser.py: A Lean Persistent Homology Library for Python}, journal = {Journal of Open Source Software} }

Please correct!
