### 2025-07-19 uislambekov/TDAvec#10 @adityapt
I'm seeing other issues too. These should be the right references as per JOSS standards:

@article{ali2023survey,
  title   = {A Survey of Vectorization Methods in Topological Data Analysis},
  author  = {Ali, Dashti and Asaad, Aras and Jimenez, Maria-Jose and Nanda, Vidit and Paluzo-Hidalgo, Eduardo and Soriano-Trigueros, Manuel},
  journal = {IEEE Transactions on Pattern Analysis and Machine Intelligence},
  volume  = {45},
  number  = {12},
  pages   = {14069--14080},
  year    = {2023},
  doi     = {10.1109/TPAMI.2023.3348912}
}

@article{vanVeen2019keplermapper,
  title   = {Kepler Mapper: A Flexible Python Implementation of the Mapper Algorithm},
  author  = {Van Veen, Hendrik Jacob and Saul, Nathaniel and Eargle, David and Mangham, Sam W.},
  journal = {Journal of Open Source Software},
  volume  = {4},
  number  = {42},
  pages   = {1315},
  year    = {2019},
  doi     = {10.21105/joss.01315}
}

@article{perea2023approximating,
  title   = {Approximating Continuous Functions on Persistence Diagrams Using Template Functions},
  author  = {Perea, Jose A. and Munch, Elizabeth and Khasawneh, Firas A.},
  journal = {Foundations of Computational Mathematics},
  volume  = {23},
  number  = {4},
  pages   = {1215--1272},
  year    = {2023},
  doi     = {10.1007/s10208-022-09567-7}
}

@article{kalivsnik2019tropical,
  title   = {Tropical Coordinates on the Space of Persistence Barcodes},
  author  = {Kali{\v{s}}nik, Sara},
  journal = {Foundations of Computational Mathematics},
  volume  = {19},
  number  = {1},
  pages   = {101--129},
  year    = {2019},
  doi     = {10.1007/s10208-018-9379-y}
}

@inproceedings{diFabio2015comparing,
  title       = {Comparing Persistence Diagrams through Complex Vectors},
  author      = {Di~Fabio, Barbara and Ferri, Massimo},
  booktitle   = {Image Analysis and Processing--ICIAP 2015},
  series      = {Lecture Notes in Computer Science},
  volume      = {9279},
  pages       = {294--305},
  year        = {2015},
  publisher   = {Springer},
  doi         = {10.1007/978-3-319-23231-7_27}
}

@Manual{kernelTDA,
  title  = {kernelTDA: Statistical Learning with Kernel for Persistence Diagrams},
  author = {Padellini, Tullia and Palini, Francesco},
  year   = {2020},
  note   = {R package version 1.0.0},
  url    = {https://CRAN.R-project.org/package=kernelTDA},
  doi    = {10.32614/CRAN.package.kernelTDA}
}

@article{tralie2018ripserpy,
  title   = {Ripser.py: A Lean Persistent Homology Library for Python},
  author  = {Tralie, Christopher and Saul, Nathaniel and Bar-On, Rann},
  journal = {Journal of Open Source Software},
  volume  = {3},
  number  = {29},
  pages   = {925},
  year    = {2018},
  doi     = {10.21105/joss.00925}
}

I got some of these online. Please fix these!
