### 2025-07-19 #8532 @adityapt
Hey @uislambekov - I see issues with other References as well that don't have DOIs. Can you please look into all those? Thanks!
