### 2025-07-19 uislambekov/TDAvec#11 @adityapt
Unsubstantiated Performance Claims in the Paper
Hi, This is regarding the review: https://github.com/openjournals/joss-reviews/issues/8532

There are some unsubstantiated performance claims in the paper without any benchmark comparisons

"To achieve higher computational efficiency, all code behind the vector summaries of TDAvec is written in C++ using the Rcpp and RcppArmadillo packages."

"The package provides efficient implementations of various vector summary statistics, leveraging C++ code for improved computational performance through the Rcpp package"

If there is no proof of these, can we remove it or add reproducible benchmarks to these claims?

Thanks!
