+++ vignettes/TDAvec_vignette.Rmd
@@ -160,6 +160,76 @@ computePersistenceImage(D,homDim=1,xSeqH1,ySeqH1,sigma)
 ```
 The code for `computePersistenceImage()` is adopted from the `pers.image()` function (available in the `kernelTDA` package) with some modifications. For example, `pers.image()` uses a one-dimensional grid for homological dimension $H_0$ regardless of the filtration type. In contrast, `computePersistenceImage()` uses a one-dimensional grid only if additionally the birth values are the same (which may not be true for some filtration types). Moreover, `pers.image()` uses a square grid (e.g., 10 by 10) for vectorization, whereas `computePersistenceImage()` is not limited to such a grid and can compute vector summaries using a rectangular grid (e.g., 10 by 20). 
 
+Finally, we compare the computational costs of two vector-based vectorization methods—`persistence landscape` and `persistence silhouette`—across three `R` packages: `TDA`, `TDAkit`, and `TDAvec`. These are the only vector summaries consistently implemented in these three packages. For persistence diagrams with sizes ranging from 100 to 1000 points, we generate random birth and death values, construct the diagrams, and apply each method using the appropriate package functions. To assess performance, we use the `microbenchmark` package to record the median execution time (in milliseconds) for each method, averaged over 10 repetitions.
+
+```{r message=FALSE, warning=FALSE}
+library(TDA)
+library(TDAkit)
+library(microbenchmark)
+
+N <- seq(100,1000,by=100) # sequence of diagram sizes
+method <- c('landscape-TDA','landscape-TDAkit','landscape-TDAvec',
+            'silhouette-TDA','silhouette-TDAkit','silhouette-TDAvec')
+cost <- matrix(ncol = length(method),nrow = length(N))
+colnames(cost) <- method; rownames(cost) <- N
+for (i in seq_along(N)){
+  n <- N[i] 
+  b <- runif(n) # birth values
+  d <- b + runif(n) # death values
+  pd <- cbind(0,b,d) # persistence diagram
+  pdList <- list(Dimension=rep(0,n),Birth=b,Death=d) # PD as a list for TDAkit
+  attr(pdList,"class") <- "homology" 
+  scaleSeq = seq(0,2,length.out=101) # sequence of scale values
+
+  mb <- summary(microbenchmark(
+      TDA::landscape(pd,dimension = 0,KK=1:5,tseq = scaleSeq),
+      TDAkit::diag2landscape(pdList,dimension = 0,k=5,nseq = 101), # nseq: grid size
+      TDAvec::computePersistenceLandscape(pd,homDim = 0,scaleSeq,k=5),
+      
+      TDA::silhouette(pd,p=1,dimension = 0,tseq = scaleSeq),
+      TDAkit::diag2silhouette(pdList,dimension = 0,p=1,nseq = 101),
+      TDAvec::computePersistenceSilhouette(pd,homDim = 0,scaleSeq,p=1,evaluate = "points"),
+      unit = 'ms', # unit: milliseconds
+      times = 10 # number of times to call each function
+  ))
+  cost[i,] <- mb[['median']]
+}
+
+print(t(cost),digits = 2)
+```
+
[…]
