### 2025-07-29 uislambekov/TDAvec#10 @uislambekov
Thank you! Our initial submission included an additional section that briefly surveyed existing R and Python tools for TDA. This section was later removed to comply with the 1000-word limit. Some of the references you mentioned were only relevant to that removed section and don't appear in the updated paper.md and paper.bib files. All other references have been fixed with added DOIs.
