+++ JOSS/paper.bib
@@ -188,6 +188,18 @@
 	doi={10.32614/CRAN.package.Rcpp}
 }
 
+  @Article{RcppArmadillo,
+    title = {RcppArmadillo: Accelerating R with high-performance C++
+      linear algebra},
+    author = {Dirk Eddelbuettel and Conrad Sanderson},
+    journal = {Computational Statistics and Data Analysis},
+    year = {2014},
+    volume = {71},
+    month = {March},
+    pages = {1054--1063},
+    doi = {10.1016/j.csda.2013.02.005},
+  }
+
 
 
 @article{perea2023approximating,
@@ -245,7 +257,7 @@ doi = {10.4310/HHA.2016.v18.n1.a21}
 }
 
 
-article{ali2023survey,
+@article{ali2023survey,
   author={Ali, Dashti and Asaad, Aras and Jimenez, Maria-Jose and Nanda, Vidit and Paluzo-Hidalgo, Eduardo and Soriano-Trigueros, Manuel},
   journal={IEEE Transactions on Pattern Analysis and Machine Intelligence},
   title={A Survey of Vectorization Methods in Topological Data Analysis},
