### 2025-07-29 uislambekov/TDAvec#11 @uislambekov
Thank you! Our original submission had elaborated on this point a biy, but that content was removed to comply with the 1000-word limit. For example, we had noted that some existing CRAN R packages rely on standard R functions to vectorize persistence diagrams, which may not scale well for large-scale computations. In contrast, all TDAvec functions are implemented in C++ to ensure higher computational efficiency.

To address your point, we have added run-time comparisons between TDAvec and other R packages to the package vignette and referenced this in the updated paper.md. Due to the word count constraint, the full analysis could not be included in the paper itself. The updated vignette is available at: https://github.com/uislambekov/TDAvec, and we will be updating the CRAN version shortly.

In summary, the run-time comparisons show that TDAvec consistently outperforms competing R packages in terms of computational efficiency.
