### 2025-07-29 uislambekov/TDAvec#10 @adityapt
Thanks! So, to avoid confusion, does it make sense to remove the references that aren't valid to the word-count limited paper you submitted?
