### 2025-07-30 uislambekov/TDAvec#10 @uislambekov
Let me clarify my point. The references were removed because they are no longer cited in the updated draft, not because they are invalid. In our original submission, we included a section that surveyed various R and Python libraries for TDA and their scopes, citing a number of sources. However, since that material did not directly support the statement of need, we removed it in order to stay within the 1000-word limit.

I had assumed that references not cited in the revised text would automatically be excluded from the generated PDF. Since they still appeared, I had to remove them manually.
