### 2025-08-03 uislambekov/TDAvec#11 @adityapt
This is nice! Tested it and I see the performance differences clearly. Thanks @uislambekov !

Just a small recommendation, as I mentioned in my previous issues as well - the vignette just assumes that some of the libraries are installed - a better way to deal them would be to use this methodology for better error handling:

stats_pkg_flag <- requireNamespace("TDAstats", quietly = TRUE)

if (stats_pkg_flag) {
library(TDAstats) # load it only if present
} else {
message("NOTE: TDAstats not installed; code chunks that rely on it will be shown but not run.")

Also, do we have any performance benchmarks for the python version of the package as well?
