### 2025-08-04 uislambekov/TDAvec#11 @uislambekov
Thank you for the suggestion. I've updated the .Rmd vignette file on Github with the proposed change.

We currently don't have the performance benchmarks for the python version of TDAvec. The python version was entirely handled by my collaborator. I will ask him to add these performance benchmarks.

Thanks again!
