### 2025-08-05 uislambekov/TDAvec#8 @adityapt
Hey @uislambekov - Just checking if there is any progress on this. I don't see the tests being setup yet. This, apart from the Python performance stats are the only things remaining to close the package review.
