+++ python/tdavec/py_times.txt
@@ -1,0 +1,12 @@
+Hello from tdavec, faster TDA vectorization!
+n, min, max, mean, median, stddev
+100, 0.000135, 0.000205, 0.000146, 0.000137, 0.000022
+200, 0.000223, 0.000310, 0.000235, 0.000226, 0.000027
+300, 0.000309, 0.000756, 0.000360, 0.000312, 0.000140
+400, 0.000646, 0.001486, 0.000777, 0.000711, 0.000252
+500, 0.000733, 0.001223, 0.000846, 0.000805, 0.000140
+600, 0.000870, 0.001201, 0.000955, 0.000943, 0.000094
+700, 0.001040, 0.001510, 0.001120, 0.001084, 0.000140
+800, 0.001145, 0.001529, 0.001235, 0.001209, 0.000108
+900, 0.001253, 0.001829, 0.001372, 0.001327, 0.000166
+1000, 0.001431, 0.002084, 0.001531, 0.001472, 0.000196
