### 2025-08-09 78ace81
better time profyling
