+++ python/tdavec/py_times.txt
@@ -1,12 +1,12 @@
 Hello from tdavec, faster TDA vectorization!
 n, min, max, mean, median, stddev
-100, 0.000135, 0.000205, 0.000146, 0.000137, 0.000022
-200, 0.000223, 0.000310, 0.000235, 0.000226, 0.000027
-300, 0.000309, 0.000756, 0.000360, 0.000312, 0.000140
-400, 0.000646, 0.001486, 0.000777, 0.000711, 0.000252
-500, 0.000733, 0.001223, 0.000846, 0.000805, 0.000140
-600, 0.000870, 0.001201, 0.000955, 0.000943, 0.000094
-700, 0.001040, 0.001510, 0.001120, 0.001084, 0.000140
-800, 0.001145, 0.001529, 0.001235, 0.001209, 0.000108
-900, 0.001253, 0.001829, 0.001372, 0.001327, 0.000166
-1000, 0.001431, 0.002084, 0.001531, 0.001472, 0.000196
+100, 0.000277, 0.001304, 0.000306, 0.000278, 0.000145
+200, 0.000403, 0.000739, 0.000444, 0.000422, 0.000060
+300, 0.000531, 0.000840, 0.000562, 0.000538, 0.000061
+400, 0.000903, 0.026450, 0.001459, 0.000921, 0.003607
+500, 0.001063, 0.001431, 0.001108, 0.001100, 0.000048
+600, 0.001159, 0.001457, 0.001181, 0.001165, 0.000045
+700, 0.001289, 0.001778, 0.001334, 0.001306, 0.000077
+800, 0.001444, 0.001868, 0.001507, 0.001492, 0.000079
+900, 0.001583, 0.001814, 0.001652, 0.001642, 0.000054
+1000, 0.001734, 0.002450, 0.001822, 0.001804, 0.000107
