### 2025-08-09 uislambekov/TDAvec#8 @uislambekov
R unit tests have been added, passing without any issues. Also, we have added Python performance stats to the TDAvec vignette.
