### 2025-08-09 b82eb17
time_test: vect_name in CLI parame, out file name selected accordingly
