### 2025-08-09 uislambekov/TDAvec#8 @adityapt
Hey @uislambekov - The tests are working fine, but I think this - '**../python/tdavec/unitCircle.csv**' may fail when the package is installed.

May be use system.file() to detect the data - move it out of python folder to another folder, and it should work.
