### 2025-08-09 uislambekov/TDAvec#11 @adityapt
Thanks for adding the python time_tests.py. I am closing this issue.
