+++ python/tdavec/__init__.py
@@ -1,5 +1,3 @@
-print("Hello from tdavec, faster TDA vectorization!")
-
 from .vectorizer import TDAvectorizer
 from .vectorizer import createEllipse, pmin, pmax, test_package
 
