### 2025-08-10 666b434
adding arma:: to all export function declarations
