+++ src/PersistentEntropy.cpp
@@ -8,7 +8,8 @@ using namespace std;
 // [[Rcpp::depends(RcppArmadillo)]]
 
 // [[Rcpp::export]]
-NumericVector computePersistentEntropy(const mat& D, const int& homDim, const vec& scaleSeq, string evaluate = "intervals") {
+NumericVector computePersistentEntropy(const arma::mat& D, const int& homDim, const arma::vec& scaleSeq,
+                                       std::string evaluate = "intervals") {
 
   // Get indices of rows where D.col(0) == homDim
   uvec indices = find(D.col(0) == homDim);
