+++ src/PersistenceSilhouette.cpp
@@ -6,7 +6,7 @@ using namespace Rcpp;
 // [[Rcpp::depends(RcppArmadillo)]]
 
 // [[Rcpp::export]]
-NumericVector computePersistenceSilhouette(const mat& D, const int& homDim, const vec& scaleSeq,
+NumericVector computePersistenceSilhouette(const arma::mat& D, const int& homDim, const arma::vec& scaleSeq,
                         const double& p=1.0, const std::string& evaluate = "intervals") {
 
   uvec indices = find(D.col(0) == homDim); // Get indices of rows where D.col(0) == homDim
