+++ src/PersistenceLandscape.cpp
@@ -7,8 +7,10 @@ using namespace std;
 // [[Rcpp::depends(RcppArmadillo)]]
 
 // [[Rcpp::export]]
-mat computePersistenceLandscape(const mat& D, const int& homDim, const vec& scaleSeq, const int& k=1,
-                                const bool& generalized=false, const string& kernel = "triangle",
+arma::mat computePersistenceLandscape(const arma::mat& D, const int& homDim,
+                                const arma::vec& scaleSeq, const int& k=1,
+                                const bool& generalized=false,
+                                const std::string& kernel = "triangle",
                                 Nullable<double> h = R_NilValue) {
 
   uvec indices = find(D.col(0) == homDim); // Get indices of rows where D.col(0) == homDim
