+++ vignettes/TDAvec_vignette.Rmd
@@ -210,7 +210,7 @@ Let us also download the results for python simulation, created using `./python/
 
 
 ```{r}
-py_data <- read.csv("../python/tdavec/benchmark_ps.csv")
+py_data <- read.csv(system.file("extdata", "benchmark_ps.csv", package = "TDAvec"))
 py_cost <- rbind(t(cost), 1e3*py_data$mean)
 rownames(py_cost)[nrow(py_cost)] <- "silhouette-python"
 print(py_cost, digits = 3)
