### 2025-08-12 #8532 @zplus11
Hi @uislambekov, the examples given in documentation (both R and Python) work as shown. Installing the Python version of TDAvec requires Microsoft Build Tools, I think that a small mention of this may be made in the python\README.md file so that users don't run into errors. With this, all of my package and documentation related points will be check-marked.

In paper.pdf, I have a few minor suggestions/corrections:&mdash;

- On page 1 line 23, "etc" should end with a full stop.
- On page 1 lines 24&ndash;25, rather than *x*-coordinate and *y*-coordinate, it may be more appropriate to directly write *b*<sub>*i*</sub> and *d*<sub>*i*</sub>.
- On page 1 lines 36 & 38, the hyphens (-) should be replaced with em dashes (&mdash;).
- On page 1, the first footnote should end with a full stop.
- On page 2 line 70, you may want to remove the commas before d*t*. I’m not sure if that’s standard practice, but in this case they create confusion about what separates the tuple elements.
- On page 2 line 71, the citation for Adam H. *et al.* should match the style of other citations. In this instance, it appears as the reference number in round brackets, whereas others use author names.
