### 2025-08-12 29ffd49
no manual install instructions
