### 2025-08-12 8057a7d
install instructions updated
