+++ python/README.md
@@ -32,10 +32,11 @@ Note that the package was tested only on python 3.12.
 
 into your environment. 
 
-You can also install the current verion from the GitHub with
-
-    pip install -e "git+https://github.com/uislambekov/TDAvec.git#egg=tdavec&subdirectory=python"
-
+If a wheel is not available for your system, pip will build from source.
+This may require a C/C++ compiler:
+* **Windows**: Microsoft Build Tools
+* **macOS**: Xcode Command Line Tools (xcode-select --install)
+* **Linux**: gcc/clang and Python development headers (python3-dev)
 
 In order to check if the intallation process was completed, you can run python and evaluate the following lines:
 
