### 2025-08-16 7b805b9
fix minor typos in python\README.md and JOSS\paper.md
