+++ python/README.md
@@ -38,7 +38,7 @@ This may require a C/C++ compiler:
 * **macOS**: Xcode Command Line Tools (xcode-select --install)
 * **Linux**: gcc/clang and Python development headers (python3-dev)
 
-In order to check if the intallation process was completed, you can run python and evaluate the following lines:
+In order to check if the installation process was completed, you can run python and evaluate the following lines:
 
     > from tdavec import test_package
     > X, D, PS = test_package()
@@ -47,7 +47,7 @@ This function will create a simple point cloud, build a persistence diagram, cac
 
 # Usage
 
-In this section some simple example of package usage is demonstrated.
+In this section, some simple examples of package usage are demonstrated.
 
 We will start with loading `TDAvec` library and some other packages:
 
@@ -59,7 +59,7 @@ We will start with loading `TDAvec` library and some other packages:
     import pandas as pd
     import numpy as np
 
-As a sample data we will work with set of point clouds, that represent defomed elipses with randomly selected squize rations:
+As a sample data we will work with set of point clouds, that represent deformed elipses with randomly selected squeeze ratios:
 
     np.random.seed(42)
     epsList = np.random.uniform(low = 0, high = 1, size = 500)
@@ -77,13 +77,13 @@ Here are some examples:
 
 ![Point clouds](https://raw.githubusercontent.com/uislambekov/TDAvec/main/python/readme_figs/clouds.png)
 
-In order to generate Persistence Diagrams one need to create TDAvectorizer object and fit fit it:
+In order to generate Persistence Diagrams one needs to create `TDAvectorizer` object and fit it:
 
     v = TDAvectorizer()
     v.setParams({"scale":np.linspace(0, 2, 10)})
     v.fit(clouds)
 
-Here are the examples of the generated persistence diagrams:
+Here are the examples of the generated persistence diagram:
 
     for i in range(4):
         plt.subplot(2,2,i+1)
@@ -97,7 +97,7 @@ Here are the examples of the generated persistence diagrams:
 
 ![PDs](https://raw.githubusercontent.com/uislambekov/TDAvec/main/python/readme_figs/PDs.png)
 
-Once TDAvectorizer object is fitted, one can calculate vectorization by calling `transorm()` method of this object:
+Once `TDAvectorizer` object is fitted, one can calculate vectorization by calling `transorm()` method of this object:
 
     X = v.transform(output="PS", homDim=1)
     for i, e in enumerate(epsList[:4]):
@@ -108,9 +108,9 @@ Once TDAvectorizer object is fitted, one can calculate vectorization by calling 
 
 ![Vectorizations](https://raw.githubusercontent.com/uislambekov/TDAvec/main/python/readme_figs/Vects.png)
 
-These vectorizations can be used as predictors for ML problem, whose goal is to predict the original deformation parameter. We will use a simple `sklearn.LinearRegression` model to solve the problem
+These vectorizations can be used as predictors for ML problem, whose goal is to predict the original deformation parameter. We will use a simple `sklearn.LinearRegression` model to solve the problem.
 
-Here is a simple function the for any given set of predictors creates the model, solves it, and retirns the results:
+Here is a simple function, that for any given set of predictors creates the model, solves it, and returns the results:
 
     def makeSim(X, y=epsList):
         Xtrain, Xtest, ytrain, ytest = train_test_split(X, y, train_size=0.8, random_state=42)
@@ -120,7 +120,7 @@ Here is a simple function the for any given set of predictors creates the model,
         res = {"method":method, "homDim":homDim, "test_preds":test_preds, "y_test":ytest, "score":score}
         return res
 
-In the loop below a systematic scan over different vectorizattion methods and homological dimensions is performed:
[…]
