+++ .github/workflows/python-package-conda.yml
@@ -32,10 +32,4 @@ jobs:
     - name: Test with pytest
       run: |
         cd ./python/
-        pytest --cov-report "xml:coverage.xml" --cov=.
-    - name: Create Coverage
-      if: ${{ github.event_name == 'pull_request' }}
-      uses: orgoro/coverage@v3
-      with:
-          coverageFile: coverage.xml
-          token: ${{ secrets.GITHUB_TOKEN }}
+        pytest
