### 2025-08-19 da7fd80
auto test using MacOS
