+++ .github/workflows/python-package-conda.yml
@@ -16,7 +16,7 @@ permissions:
 jobs:
   build:
 
-    runs-on: ubuntu-latest
+    runs-on: macos-latest
 
     steps:
     - uses: actions/checkout@v4
@@ -29,7 +29,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install tdavec
         pip install pytest
+        pip install pyprojroot
     - name: Test with pytest
       run: |
-        cd ./python/
         pytest
