+++ .github/workflows/python-package-conda.yml
@@ -27,9 +27,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install tdavec
-        pip install pytest
-        pip install pyprojroot
+        pip install tdavec pytest pyprojroot
     - name: Test with pytest
       run: |
-        pytest
+        python python/tdavec/unit_test.py
