+++ .github/workflows/R_test.yml
@@ -1,0 +1,39 @@
+name: Test
+
+on:
+  push:
+    branches: [ "main" ]
+  pull_request:
+    branches: [ "main" ]
+
+jobs:
+  build:
+    runs-on: macos-latest
+
+    strategy:
+      matrix:
+        r: [ 'release' ]
+
+    steps:
+      - uses: actions/checkout@v3
+
+      - uses: r-lib/actions/setup-r@v2
+        with:
+          r-version: ${{ matrix.r }}
+
+      - uses: r-lib/actions/setup-r-dependencies@v2
+        with:
+          cache-version: 2
+          extra-packages: |
+            any::tidyverse
+            any::knitr
+            any::devtools
+            any::testthat
+
+      - name: Install from CRAN
+        run: |
+          Rscript -e "install.packages('TDAvec', repos='https://cran.r-project.org')"
+
+      - name: Run tests
+        run: |
+          Rscript -e "options(rgl.useNULL=TRUE); devtools::test()"
