### 2025-09-20 #8532 @ujjwalkarn
@uislambekov @ALuchinsky As a next step, could you please:

1. Double check if DOIs can be found for the 4 references which do not have it as of now.
2. Make a tagged release of your software, and list the version tag of the archived version here.
3. Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
4. Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
5. Please list the DOI of the archived version here.
