+++ paper.bib
@@ -20,6 +20,7 @@
   pages={77--102},
   year={2015},
   publisher={JMLR. org}
+  doi={10.5555/2789272.2789275}
 }
 
 @inproceedings{chazal2014stochastic,
@@ -40,6 +41,7 @@
   pages={218--252},
   year={2017},
   publisher={JMLR. org}
+  doi={10.5555/3122009.3122017}
 }
 
 @article{Carlsson:2009,
@@ -146,6 +148,7 @@
 	number={268},
 	pages={1--33},
 	year={2022}
+	doi={10.5555/3586589.3586857}
 }
 
 @incollection{nanda2013simplicial,
