### 2025-09-22 fbb3b32
python tests on macos-15
