+++ .github/workflows/python-package-conda.yml
@@ -16,7 +16,7 @@ permissions:
 jobs:
   build:
 
-    runs-on: macos-latest
+    runs-on: macos-15
 
     steps:
     - uses: actions/checkout@v4
