### 2025-09-23 ff4a6b6
resolving ripser instability
