### 2025-10-18 #8532 @crvernon
@uislambekov and @ALuchinsky - I need you to address the following before I accept this submission for publication:

In your archive:
- [ ] The name and author list (including the order) of your Zenodo archive needs to match the title and authors of your paper exactly.  Please edit the metadata in your Zenodo archive to reflect this.

In your paper:
- [ ] LINE 119:  The "r" and "c" in "r and c++" should be capitalized.  You can do this in your bib file by wrapping any characters you wish to maintain the formatting of in curly brackets.
- [ ] LINE 140:  The "e" in "euler" should be capitalized.

Let me know when these changes have been made and I will accept.  Thank you.
