+++ paper.bib
@@ -130,7 +130,7 @@
 }
 
 @article{richardson2014efficient,
-	title={Efficient classification using the Euler characteristic},
+	title={Efficient classification using the {E}uler characteristic},
 	author={Richardson, Eitan and Werman, Michael},
 	journal={Pattern Recognition Letters},
 	volume={49},
@@ -181,7 +181,7 @@
 
 
   @Manual{Rcpp,
-	title = {Rcpp: Seamless R and C++ Integration},
+	title = {Rcpp: Seamless {R} and {C++} Integration},
 	author = {Dirk Eddelbuettel and Romain Francois and JJ Allaire and
 	Kevin Ushey and Qiang Kou and Nathan Russell and Inaki Ucar and
 	Douglas Bates and John Chambers},
