### 2025-10-18 2c675e7
capital letters in JOSS/paper.bib
