+++ JOSS/paper.bib
@@ -19,7 +19,7 @@
   number={1},
   pages={77--102},
   year={2015},
-  publisher={JMLR. org}
+  publisher={JMLR. org},
   doi={10.5555/2789272.2789275}
 }
 
@@ -40,7 +40,7 @@
   number={1},
   pages={218--252},
   year={2017},
-  publisher={JMLR. org}
+  publisher={JMLR. org},
   doi={10.5555/3122009.3122017}
 }
 
@@ -147,7 +147,7 @@
 	volume={23},
 	number={268},
 	pages={1--33},
-	year={2022}
+	year={2022},
 	doi={10.5555/3586589.3586857}
 }
 
