### 2025-10-18 #8532 @ALuchinsky
Bibliography issues are fixed too
