### 2025-10-20 #8532 @uislambekov
We would like to thank JOSS for considering our submission, and the editor and reviewers for their diligent work.
